requirements.txt
pycairo==1.19.1
numpy==1.18.1
